add and remove stars
