+++ src/main.rs
@@ -1,8 +1,11 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")] // hide console window on Windows in release
 
-use std::cmp::max;
+use std::f32::consts::TAU;
+use rand::prelude::*;
 
-use eframe::{egui::{self, Ui, Sense, Id, PointerButton}, epaint::{Color32, Pos2, Vec2, Stroke, Rect}};
+use eframe::{egui::{self, Ui, Sense, Id}, epaint::{Color32, Pos2, Vec2, Stroke, Rect}};
+use gravity::Simulation;
+use gravity;
 
 fn main() -> Result<(), eframe::Error> {
     let options = eframe::NativeOptions {
@@ -16,16 +19,6 @@ fn main() -> Result<(), eframe::Error> {
     )
 }
 
-struct Field {
-
-}
-
-impl Field {
-    fn gradient(p: &Pos2) -> Vec2 {
-        Vec2::new(0.0, 0.0)
-    }
-}
-
 #[derive(PartialEq)]
 enum Mode {
     Add,
@@ -33,7 +26,8 @@ enum Mode {
 }
 
 struct GravityApp {
-    field: Field,
+    rng: rand::rngs::ThreadRng,
+    simulation: Simulation,
     mode: Mode,
     radius: f32,
     orbital_velocity: bool,
@@ -42,7 +36,8 @@ struct GravityApp {
 impl Default for GravityApp {
     fn default() -> Self {
         Self {
-            field: Field {  },
+            rng: rand::thread_rng(),
+            simulation: Simulation::new(),
             mode: Mode::Add,
             radius: 64.0,
             orbital_velocity: true,
@@ -50,8 +45,10 @@ impl Default for GravityApp {
     }
 }
 
-fn draw(ui: &mut Ui, field: &Field, spacing: Vec2) {
+fn draw(ui: &mut Ui, simulation: &Simulation, spacing: Vec2) {
     let size = ui.available_size();
+    
+    // draw grid
     let color = Color32::from_rgb(00, 0x8b, 0x8b);
     let stroke = Stroke { width: 1.0, color};
     let grid_size = size / spacing;
@@ -66,6 +63,22 @@ fn draw(ui: &mut Ui, field: &Field, spacing: Vec2) {
             painter.line_segment([p, p + d01], stroke);
         }
     }
+
+    // draw masses
+    let color = Color32::from_rgba_premultiplied(00, 0x8b, 0x8b, 0x40);
+    for position in simulation.state.positions.iter() {
+        painter.circle_filled(Pos2::new(position.x, position.y), 3.0, color);
+    }
+}
+
+fn random_point_in_circle(rng: &mut ThreadRng, radius: f32) -> Vec2 {
+    let r = radius * rng.gen::<f32>().sqrt();
+    let theta = rng.gen::<f32>() * TAU;
+    Vec2::new(r * theta.cos(), r * theta.sin())
+}
+
+fn convert(v: Vec2) -> gravity::Vec2<f32> {
+    gravity::Vec2 { x: v.x, y: v.y }
 }
 
 impl eframe::App for GravityApp {
@@ -82,7 +95,7 @@ impl eframe::App for GravityApp {
             }
         });
         egui::CentralPanel::default().show(ctx, |ui| {
-            draw(ui, &self.field, spacing);
+            draw(ui, &self.simulation, spacing);
             let rect = Rect::from_min_size(Pos2::ZERO, ui.available_size());
             
             ui.input(|i| self.radius = (self.radius + 0.1 * i.scroll_delta.y).clamp(0.0, 1024.0));
@@ -90,13 +103,33 @@ impl eframe::App for GravityApp {
             let response = ui.interact(rect, id, Sense::hover());
             if let Some(hover_pos) = response.hover_pos() {
                 let stroke = Stroke { width: 1.0, color: Color32::WHITE};
-                ui.painter().circle(hover_pos, self.radius, Color32::TRANSPARENT, stroke);
+                ui.painter().circle_stroke(hover_pos, self.radius, stroke);
             }
             
-            let response = ui.interact(rect, id, Sense::click());
-            if response.clicked() {
+            let response = ui.interact(rect, id, Sense::click_and_drag());
+            if response.dragged() {
                 if let Some(pointer_pos) = response.interact_pointer_pos() {
-                    println!("{:?}", pointer_pos);
+                    match self.mode {
+                        Mode::Add => {
+                            let position = random_point_in_circle(&mut self.rng, self.radius) + pointer_pos.to_vec2();
+                            let velocity = gravity::Vec2::zero();
+                            self.simulation.add(&convert(position), &velocity, 1.0);
+                        },
+                        Mode::Remove => {
+                            let center = convert(pointer_pos.to_vec2());
+                            let r2 = self.radius * self.radius;
+                            let mut indices = self.simulation.state.positions
+                                .iter()
+                                .enumerate()
+                                .filter(|(_, &p)| p.sub(&center).norm2() < r2)
+                                .map(|(index, _)| index)
+                                .collect::<Vec<_>>();
+                            indices.reverse();
+                            for index in indices {
+                                self.simulation.remove(index);
+                            }
+                        }                            
+                    }
                 }
             }
         });
